Resume Claims Question Generation

+++ PrepPilot-AI/src/main/java/com/example/PrepPilot/AI/dto/VerificationQuestionResponse.java
@@ -1,7 +1,7 @@
 package com.example.PrepPilot.AI.dto;
 
 public record VerificationQuestionResponse(
-        int questionId,
+        Long questionId,
         Long claimId,
         String question,
         String claim
